@@ -1,15 +1,10 @@
-import preprocess from 'svelte-preprocess';
-import { mdsvex } from 'mdsvex';
-import mdsvexConfig from './mdsvex.config.js';
-import adapter from '@sveltejs/adapter-vercel';
-
+import adapter from '@sveltejs/adapter-auto';
+import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
-  extensions: ['.svelte', ...mdsvexConfig.extensions],
-  preprocess: [mdsvex(mdsvexConfig), preprocess({ postcss: true })],
-  kit: {
-    adapter: adapter({}),
-  },
+	kit: {
+		adapter: adapter()
+	},
+	preprocess: vitePreprocess()
 };
-
 export default config;
